+++ src/routes/User.js
@@ -1,10 +1,9 @@
-/* eslint-disable max-lines-per-function */
 const passport = require("passport");
-const { body, param } = require("express-validator");
+const { body, param, query } = require("express-validator");
 const User = require("../controllers/User");
+const { createAccountLimiter, loginLimiter, basicLimiter } = require("../helpers/constants/Route");
 
 module.exports = app => {
-
     /**
      * @apiDefine JWT Bearer Authorization with JSON Web Token.
      */
@@ -15,8 +14,8 @@ module.exports = app => {
 
     /**
      * @apiDefine UserRequestBody
-     * @apiParam (Request Body Parameters) {String} email User's email.
-     * @apiParam (Request Body Parameters) {String{>=5}} password User's password.
+     * @apiParam (Request Body Parameters) {String{>=30}} email User's email.
+     * @apiParam (Request Body Parameters) {String{>= 5 && <= 50}} password User's password.
      */
 
     /**
@@ -33,9 +32,13 @@ module.exports = app => {
      * @apiGroup Users 👤
      *
      * @apiPermission Basic
+     * @apiParam (Query String Parameters) {String} [fields] Specifies which user fields to include. Each value must be separated by a `,`.
      * @apiUse UserResponse
      */
-    app.get("/users", passport.authenticate("basic", { session: false }), User.getUsers);
+    app.get("/users", passport.authenticate("basic", { session: false }), [
+        query("fields")
+            .optional(),
+    ], User.getUsers);
 
     /**
      * @api {GET} /users/:id B] Get an user
@@ -49,7 +52,7 @@ module.exports = app => {
      * - `Basic Auth`: Any user can be retrieved.
      * @apiUse UserResponse
      */
-    app.get("/users/:id", passport.authenticate(["jwt", "basic"], { session: false }), [
+    app.get("/users/:id", basicLimiter, passport.authenticate(["jwt", "basic"], { session: false }), [
         param("id")
             .isMongoId().withMessage("Must be a valid MongoId"),
     ], User.getUser);
@@ -62,13 +65,14 @@ module.exports = app => {
      * @apiUse UserRequestBody
      * @apiUse UserResponse
      */
-    app.post("/users", [
+    app.post("/users", createAccountLimiter, [
         body("email")
             .isEmail().withMessage("Must be a valid email")
-            .trim(),
+            .trim()
+            .isLength({ max: 30 }).withMessage("Can't exceed 30 characters long"),
         body("password")
             .isString().withMessage("Must be a string")
-            .isLength({ min: 5 }).withMessage("Must be at least 5 characters long"),
+            .isLength({ min: 5, max: 50 }).withMessage("Must be at least 5 characters long"),
     ], User.postUser);
 
     /**
@@ -79,12 +83,13 @@ module.exports = app => {
      * @apiUse UserRequestBody
      * @apiSuccess {String} token JSON Web Token to keep for further route authentication.
      */
-    app.post("/users/login", [
+    app.post("/users/login", loginLimiter, [
         body("email")
             .isEmail().withMessage("Must be a valid email")
-            .trim(),
+            .trim()
+            .isLength({ max: 30 }).withMessage("Can't exceed 30 characters long"),
         body("password")
             .isString().withMessage("Must be a string")
-            .isLength({ min: 5 }).withMessage("Must be at least 5 characters long"),
+            .isLength({ min: 5, max: 50 }).withMessage("Must be at least 5 characters long"),
     ], User.login);
 };