+++ tests/e2e/game/raven-marked-dead-player-game.test.js
@@ -9,15 +9,14 @@ chai.use(chaiHttp);
 const { expect } = chai;
 
 const credentials = { email: "test@test.fr", password: "secret" };
-const players = [
-    { name: "Dug", role: "raven" }, // O
-    { name: "Dyg", role: "villager" }, // 1
-    { name: "Deg", role: "werewolf" }, // 2
-    { name: "Dog", role: "werewolf" }, // 3
+let players = [
+    { name: "Dug", role: "raven" },
+    { name: "Dyg", role: "villager" },
+    { name: "Deg", role: "werewolf" },
+    { name: "Dog", role: "werewolf" },
 ];
 let token, game;
 
-// eslint-disable-next-line max-lines-per-function
 describe("F - Game where raven marks a player who dies during the night", () => {
     before(done => resetDatabase(done));
     after(done => resetDatabase(done));
@@ -45,7 +44,7 @@ describe("F - Game where raven marks a player who dies during the night", () => 
     it("🎲 Creates game with JWT auth (POST /games)", done => {
         chai.request(app)
             .post("/games")
-            .set({ "Authorization": `Bearer ${token}` })
+            .set({ Authorization: `Bearer ${token}` })
             .send({ players })
             .end((err, res) => {
                 expect(res).to.have.status(200);
@@ -54,15 +53,17 @@ describe("F - Game where raven marks a player who dies during the night", () => 
             });
     });
     it("👪 All elect the raven as the sheriff (POST /games/:id/play)", done => {
-        const { players } = game;
+        players = game.players;
         chai.request(app)
             .post(`/games/${game._id}/play`)
-            .set({ "Authorization": `Bearer ${token}` })
-            .send({ source: "all", action: "elect-sheriff", votes: [
-                { from: players[0]._id, for: players[1]._id },
-                { from: players[1]._id, for: players[0]._id },
-                { from: players[2]._id, for: players[0]._id },
-            ] })
+            .set({ Authorization: `Bearer ${token}` })
+            .send({
+                source: "all", action: "elect-sheriff", votes: [
+                    { from: players[0]._id, for: players[1]._id },
+                    { from: players[1]._id, for: players[0]._id },
+                    { from: players[2]._id, for: players[0]._id },
+                ],
+            })
             .end((err, res) => {
                 expect(res).to.have.status(200);
                 game = res.body;
@@ -76,13 +77,11 @@ describe("F - Game where raven marks a player who dies during the night", () => 
             });
     });
     it("🐺 Werewolves eat the villager (POST /games/:id/play)", done => {
-        const { players } = game;
+        players = game.players;
         chai.request(app)
             .post(`/games/${game._id}/play`)
-            .set({ "Authorization": `Bearer ${token}` })
-            .send({ source: "werewolves", action: "eat", targets: [
-                { player: players[1]._id },
-            ] })
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "werewolves", action: "eat", targets: [{ player: players[1]._id }] })
             .end((err, res) => {
                 expect(res).to.have.status(200);
                 game = res.body;
@@ -93,13 +92,11 @@ describe("F - Game where raven marks a player who dies during the night", () => 
             });
     });
     it("🐦 Raven marks the villager (POST /games/:id/play)", done => {
-        const { players } = game;
+        players = game.players;
         chai.request(app)
             .post(`/games/${game._id}/play`)
-            .set({ "Authorization": `Bearer ${token}` })
-            .send({ source: "raven", action: "mark", targets: [
-                { player: players[1]._id },
-            ] })
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "raven", action: "mark", targets: [{ player: players[1]._id }] })
             .end((err, res) => {
                 expect(res).to.have.status(200);
                 game = res.body;
@@ -116,13 +113,11 @@ describe("F - Game where raven marks a player who dies during the night", () => 
         done();
     });
     it("👪 One vote only for raven, dead villager is not nominated despite the fact he has two votes (POST /games/:id/play)", done => {
-        const { players } = game;
+        players = game.players;
         chai.request(app)
             .post(`/games/${game._id}/play`)
-            .set({ "Authorization": `Bearer ${token}` })
-            .send({ source: "all", action: "vote", votes: [
-                { from: players[2]._id, for: players[0]._id },
-            ] })
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "all", action: "vote", votes: [{ from: players[2]._id, for: players[0]._id }] })
             .end((err, res) => {
                 expect(res).to.have.status(200);
                 game = res.body;